Replace order minion with task minion and refactor base minion
