+++ lib/catalog/minion.rb
@@ -1,5 +1,5 @@
 require "catalog/api/minion/approval"
-require "catalog/api/minion/order"
+require "catalog/api/minion/task"
 require "catalog/api/minion/logging"
 require "catalog/api/minion/version"
 require "catalog/api/minion/base"
